Schemas available outside of tests
