+++ tests/fixtures.py
@@ -1,10 +1,6 @@
-import json
-import pathlib
-from importlib.resources import files
-
 import pytest
 
-from formkit_ninja import schemas as schema_path
+from formkit_ninja.schemas import Schemas
 
 """
 These are the forms used in Partisipa
@@ -22,28 +18,6 @@ for form_name, form_defn in _run_tsx().items():
 
 
 """
-
-
-class Schemas:
-    def __init__(self):
-        # All json files
-        schema_files: list[pathlib.Path] = list(files(schema_path).glob("*.json"))
-        # Name of the schema: path
-        schemas = {path_.name[:-5]: path_ for path_ in schema_files}
-        self.schemas = schemas
-
-    def list_schemas(self) -> list[str]:
-        return schemas.keys()
-
-    def as_text(self, schema: str):
-        return self.schemas[schema].read_text()
-
-    def as_json(self, schema: str):
-        return json.loads(self.as_text(schema))
-
-    def as_dict(self, schema: str):
-        return json.dumps(self.as_json(schema))
-
 
 schemas = Schemas()
